fix: Source selection had 2 bugs (#52)
Fixes #53

I noticed that the tests had random failures and so I went over them
with a fine-toothed comb. I found that the tests still incorrectly
asserted on the source being `samples` in some places, and so 2 bugs
with source selection got through. This fixes all three of those issues.