fix: Fix duplicate logic (#56)
Fixes #54 
Fixes #55

The updated logic now only rejects VDSs if there is a successful VDS
already submitted for that commit by that team, instead of _any_ VDS.
And it never rejects duplicate GPs, it only logs them as duplicates in
the audit log.

As part of this, I had to handle what happens if a VDS comes in for a
commit that currently has a VDS in `PENDING` (from the same team). I
added a postgres-backed lock to prevent multiple VDSs from being tested
at the same time, producing a lame implementation of a job queue. Before
competition time, we'll have to factor these jobs out into a different
system so they're not just running as async tasks, so this will probably
be removed, but it will work for now.

In order to use the sqlalchemy distributed lock library I picked, I had
to use stock sqlalchemy. At the original implementation time I used
AIOPG to get async postgres connections and so I had to do some funky
stuff to integrate it with sqlalchemy.

Turns out sqlalchemy has a standard asyncio postgres backend based on
asyncpg, which means we can use the distributed lock library and
asynchronous postgres, but I had to do the busywork of converting from
aiopg to stock sqlalchemy. That accounts for most of the code changes
here. I've called out the changes I wasn't forced into, and you can see
that the pytest tests & e2e test are passing.