Create redis creds on prestart (#110)
Moving this responsibility to cAPI, which also issues the SAS URLs,
since Redis' state is planned to have the same lifetime as cAPI's.

I am planning on pregenerating users & passwords for all workers and
putting them in one k8s secret. I'll mount the whole secret as
`/etc/capi/workers`, and use each key in `envFrom` on the workers
themselves.